+++ lib/compat/wordpress-6.9/class-gutenberg-rest-static-templates-controller.php
@@ -2,7 +2,7 @@
 
 class Gutenberg_REST_Static_Templates_Controller extends WP_REST_Templates_Controller {
 	public function __construct() {
-		$this->rest_base = 'wp_registered_template';
+		$this->rest_base = 'registered-templates';
 		$this->namespace = 'wp/v2';
 	}