+++ lib/compat/wordpress-6.9/command-palette.php
@@ -8,7 +8,9 @@
 function gutenberg_enqueue_command_palette_assets() {
 	global $menu, $submenu;
 
-	$command_palette_settings = array();
+	$command_palette_settings = array(
+		'is_network_admin' => is_network_admin(),
+	);
 
 	if ( $menu ) {
 		$menu_commands = array();