+++ packages/block-editor/src/components/global-styles/index.js
@@ -1,5 +1,4 @@
 export { useSettingsForBlockElement } from './hooks';
-export { GlobalStylesContext } from './context';
 export {
 	default as TypographyPanel,
 	useHasTypographyPanel,
@@ -24,5 +23,3 @@ export {
 	default as BackgroundPanel,
 	useHasBackgroundPanel,
 } from './background-panel';
-export { areGlobalStyleConfigsEqual } from './utils';
-export { default as getGlobalStylesChanges } from './get-global-styles-changes';