+++ packages/edit-site/src/components/page-templates/index.js
@@ -39,7 +39,8 @@ import {
 } from './fields';
 import { defaultLayouts, getDefaultView } from './view-utils';
 
-const { usePostActions, templateTitleField } = unlock( editorPrivateApis );
+const { usePostActions, usePostFields, templateTitleField } =
+	unlock( editorPrivateApis );
 const { useHistory, useLocation } = unlock( routerPrivateApis );
 const { useEntityRecordsWithPermissions } = unlock( corePrivateApis );
 
@@ -93,39 +94,32 @@ export default function PageTemplates() {
 			combinedTemplates: false,
 		} );
 	const { records: staticRecords, isResolving: isLoadingStaticData } =
-		useEntityRecordsWithPermissions( 'root', 'registeredTemplate' );
+		useEntityRecordsWithPermissions( 'root', 'registeredTemplate', {
+			// This should not be needed, the endpoint returns all registered
+			// templates, but it's not possible right now to turn off pagination
+			// for entity configs.
+			per_page: -1,
+		} );
 
 	const activeTemplates = useMemo( () => {
-		const _active = [ ...staticRecords ].filter(
-			( record ) => ! record.is_custom
-		);
+		const _active = [ ...staticRecords ];
 		if ( activeTemplatesOption ) {
 			for ( const activeSlug in activeTemplatesOption ) {
 				const activeId = activeTemplatesOption[ activeSlug ];
-				if ( activeId === false ) {
-					// Remove the template from the array.
+				// Replace the template in the array.
+				const template = userRecords.find(
+					( userRecord ) =>
+						userRecord.id === activeId &&
+						userRecord.theme === activeTheme.stylesheet
+				);
+				if ( template ) {
 					const index = _active.findIndex(
-						( template ) => template.slug === activeSlug
+						( { slug } ) => slug === template.slug
 					);
 					if ( index !== -1 ) {
-						_active.splice( index, 1 );
-					}
-				} else {
-					// Replace the template in the array.
-					const template = userRecords.find(
-						( userRecord ) =>
-							userRecord.id === activeId &&
-							userRecord.theme === activeTheme.stylesheet
-					);
-					if ( template ) {
-						const index = _active.findIndex(
-							( { slug } ) => slug === template.slug
-						);
-						if ( index !== -1 ) {
-							_active[ index ] = template;
-						} else {
-							_active.push( template );
-						}
+						_active[ index ] = template;
+					} else {
+						_active.push( template );
 					}
 				}
 			}
@@ -133,33 +127,55 @@ export default function PageTemplates() {
 		return _active;
 	}, [ userRecords, staticRecords, activeTemplatesOption, activeTheme ] );
 
-	let _records;
 	let isLoadingData;
 	if ( activeView === 'active' ) {
-		_records = activeTemplates;
 		isLoadingData = isLoadingUserRecords || isLoadingStaticData;
 	} else if ( activeView === 'user' ) {
-		_records = userRecords;
 		isLoadingData = isLoadingUserRecords;
 	} else {
-		_records = staticRecords;
 		isLoadingData = isLoadingStaticData;
 	}
 
 	const records = useMemo( () => {
+		function isCustom( record ) {
+			// For registered templates, the is_custom field is defined.
+			return (
+				record.is_custom ??
+				// For user templates it's custom if the is_wp_suggestion meta
+				// field is not set and the slug is not found in the default
+				// template types.
+				( ! record.meta?.is_wp_suggestion &&
+					! defaultTemplateTypes.some(
+						( type ) => type.slug === record.slug
+					) )
+			);
+		}
+
+		let _records;
+		if ( activeView === 'active' ) {
+			// Don't show active custom templates in the active view.
+			_records = activeTemplates.filter(
+				( record ) => ! isCustom( record )
+			);
+		} else if ( activeView === 'user' ) {
+			_records = userRecords;
+		} else {
+			_records = staticRecords;
+		}
 		return _records.map( ( record ) => ( {
 			...record,
-			_isActive: !! activeTemplates.find(
+			_isActive: activeTemplates.some(
 				( template ) => template.id === record.id
 			),
-			_isCustom:
-				record.is_custom ??
-				( ! record.meta?.is_wp_suggestion &&
-					! defaultTemplateTypes.find(
-						( type ) => type.slug === record.slug
-					) ),
+			_isCustom: isCustom( record ),
 		} ) );
-	}, [ _records, activeTemplates, defaultTemplateTypes ] );
+	}, [
+		activeTemplates,
+		defaultTemplateTypes,
+		userRecords,
+		staticRecords,
+		activeView,
+	] );
 
 	const users = useSelect(
 		( select ) => {
@@ -195,6 +211,10 @@ export default function PageTemplates() {
 		[ history, path, view?.type ]
 	);
 
+	const postTypeFields = usePostFields( {
+		postType: TEMPLATE_POST_TYPE,
+	} );
+	const dateField = postTypeFields.find( ( field ) => field.id === 'date' );
 	const themeField = useThemeField();
 	const fields = useMemo( () => {
 		const _fields = [
@@ -206,6 +226,9 @@ export default function PageTemplates() {
 		];
 		if ( activeView === 'user' ) {
 			_fields.push( themeField );
+			if ( dateField ) {
+				_fields.push( dateField );
+			}
 		}
 		const elements = [];
 		for ( const author in users ) {
@@ -219,7 +242,7 @@ export default function PageTemplates() {
 			elements,
 		} );
 		return _fields;
-	}, [ users, activeView, themeField ] );
+	}, [ users, activeView, themeField, dateField ] );
 
 	const { data, paginationInfo } = useMemo( () => {
 		return filterSortAndPaginate( records, view, fields );
@@ -236,6 +259,7 @@ export default function PageTemplates() {
 							typeof newItem.title === 'string'
 								? newItem.title
 								: newItem.title?.rendered;
+						history.navigate( `/template?activeView=user` );
 						createSuccessNotice(
 							sprintf(
 								// translators: %s: Title of the created post or template, e.g: "Hello world".
@@ -343,11 +367,6 @@ export default function PageTemplates() {
 					<duplicateAction.RenderModal
 						items={ [ selectedRegisteredTemplate ] }
 						closeModal={ () => setSelectedRegisteredTemplate() }
-						onActionPerformed={ ( [ item ] ) => {
-							history.navigate(
-								`/${ item.type }/${ item.id }?canvas=edit`
-							);
-						} }
 					/>
 				</Modal>
 			) }