+++ packages/sync/src/manager.ts
@@ -10,6 +10,7 @@ import {
 	CRDT_RECORD_MAP_KEY as RECORD_KEY,
 	LOCAL_SYNC_MANAGER_ORIGIN,
 } from './config';
+import { createPersistedCRDTDoc, getPersistedCrdtDoc } from './persistence';
 import { getProviderCreators } from './providers';
 import type {
 	CRDTDoc,
@@ -117,9 +118,22 @@ export function createSyncManager(): SyncManager {
 		// Attach observers.
 		recordMap.observeDeep( onRecordUpdate );
 
-		ydoc.transact( () => {
-			syncConfig.applyChangesToCRDTDoc( ydoc, record );
-		}, LOCAL_SYNC_MANAGER_ORIGIN );
+		// Get and apply the persisted CRDT document, if it exists.
+		const isInvalid = applyPersistedCrdtDoc( syncConfig, ydoc, record );
+
+		// If there is no persisted document or if it has been invalidated by out-of-
+		// band updates, apply changes from the current entity record to the CRDT
+		// document. This ensures that the CRDT document reflects the latest state of
+		// the entity record.
+		if ( isInvalid ) {
+			ydoc.transact( () => {
+				syncConfig.applyChangesToCRDTDoc( ydoc, record );
+			}, LOCAL_SYNC_MANAGER_ORIGIN );
+
+			const meta = createEntityMeta( objectType, objectId );
+			handlers.editRecord( { meta } );
+			handlers.saveRecord();
+		}
 	}
 
 	/**
@@ -143,6 +157,55 @@ export function createSyncManager(): SyncManager {
 		objectId: ObjectID
 	): EntityID {
 		return `${ objectType }_${ objectId }`;
+	}
+
+	/**
+	 * Apply a persisted CRDT document to the current document, if it exists.
+	 * Return true if the document exists and is valid, otherwise false. Returning
+	 * a boolean allows us to destroy the temporary document and prevent it from
+	 * leaking out.
+	 *
+	 * @param {SyncConfig} syncConfig Sync configuration for the object type.
+	 * @param {CRDTDoc}    targetDoc  Target CRDT doc.
+	 * @param {ObjectData} record     Entity record representing this object type.
+	 * @return {boolean} Whether the persisted document is non-existent or invalid.
+	 */
+	function applyPersistedCrdtDoc(
+		syncConfig: SyncConfig,
+		targetDoc: CRDTDoc,
+		record: ObjectData
+	): boolean {
+		if ( ! syncConfig.supports?.crdtPersistence ) {
+			return true;
+		}
+
+		// Get the persisted CRDT document, if it exists.
+		const tempDoc = getPersistedCrdtDoc( record );
+
+		if ( ! tempDoc ) {
+			return true;
+		}
+
+		// Apply the persisted document to the current document as a singular update.
+		// This is done even if the persisted document has been invalidated. This
+		// prevents a newly joining peer (or refreshing user) from re-initializing
+		// the CRDT document (the "initialization problem").
+		const update = Y.encodeStateAsUpdateV2( tempDoc );
+		targetDoc.transact( () => {
+			Y.applyUpdateV2( targetDoc, update );
+		}, LOCAL_SYNC_MANAGER_ORIGIN );
+
+		// Check if the persisted doc has been invalidated by out-of-band updates
+		// (e.g., a WP CLI command or direct database update) by determining if the
+		// persisted document introduces any changes that are not present in the
+		// current record. If it has been invalidated, then we return true as a
+		// signal that we need to apply the entity record to the target document.
+		const changes = syncConfig.getChangesFromCRDTDoc( tempDoc, record );
+
+		// Destroy the temporary document to prevent leaks.
+		tempDoc.destroy();
+
+		return Object.keys( changes ).length > 0;
 	}
 
 	/**
@@ -200,6 +263,10 @@ export function createSyncManager(): SyncManager {
 			await handlers.getEditedRecord()
 		);
 
+		if ( 0 === Object.keys( changes ).length ) {
+			return;
+		}
+
 		// This is a good spot to debug to see which changes are being synced. Note
 		// that `blocks` will always appear in the changes, but will only result
 		// in an update to the store if the blocks have changed.
@@ -207,7 +274,28 @@ export function createSyncManager(): SyncManager {
 		handlers.editRecord( changes );
 	}
 
+	/**
+	 * Create object meta to persist the CRDT document in the entity record.
+	 *
+	 * @param {ObjectType} objectType Object type.
+	 * @param {ObjectID}   objectId   Object ID.
+	 */
+	function createEntityMeta(
+		objectType: ObjectType,
+		objectId: ObjectID
+	): Record< string, string > {
+		const entityId = getEntityId( objectType, objectId );
+		const entityState = entityStates.get( entityId );
+
+		if ( ! entityState?.syncConfig.supports?.crdtPersistence ) {
+			return {};
+		}
+
+		return createPersistedCRDTDoc( entityState.ydoc );
+	}
+
 	return {
+		createMeta: createEntityMeta,
 		load: loadEntity,
 		unload: unloadEntity,
 		update: updateCRDTDoc,